works, need input on design.
