+++ marketplace/ui/src/components/Order/index.js
@@ -1,9 +1,8 @@
-import { Input, Tabs, Typography, DatePicker } from "antd";
-import React, { useState } from "react";
+import { Input, Tabs} from "antd";
+import React from "react";
 import { useLocation } from "react-router-dom";
 import SoldOrdersTable from "./SoldOrdersTable";
 import BoughtOrdersTable from "./BoughtOrdersTable";
-import dayjs from "dayjs";
 
 const { Search } = Input;
 
@@ -13,14 +12,8 @@ const Order = ({ user }) => {
   const onChange = (key) => {
    
   };
-  
-  const [selectedDate, setSelectedDate] = useState("");
-  const { Text } = Typography;
   const { state } = useLocation();
 
-  const onDateChange = (date) => {
-    setSelectedDate(date);
-  };
 
   return (
     <div>
@@ -28,36 +21,16 @@ const Order = ({ user }) => {
         className="mx-16 mt-14"
         defaultActiveKey={state == null ? "Sold" : state.defaultKey}
         onChange={onChange}
-        tabBarExtraContent={              
-          <div className="text-xs flex items-center">
-            <Text className="block text-primaryC text-[13px] mr-2">
-              SEARCH BY DATE
-            </Text>
-            <DatePicker
-              value={
-                selectedDate
-              }
-              disabledDate={(current) => {
-                const currentDate = dayjs().startOf('day'); // Get the start of today
-                const selectedDate = dayjs(current).startOf('day');
-
-                return selectedDate.isAfter(currentDate);
-              }}
-              onChange={onDateChange}
-              disabled={false}
-            />
-          </div>
-        }
         items={[
           {
             label: <p id="sold-tab" className="font-medium text-base">Orders (Sold)</p>,
             key: "Sold",
-            children: <SoldOrdersTable user={user} selectedDate={dayjs(selectedDate).startOf('day').unix()}/>,
+            children: <SoldOrdersTable user={user}/>,
           },
           {
             label: <p id="bought-tab" className="font-medium text-base">Orders (Bought)</p>,
             key: "Bought",
-            children: <BoughtOrdersTable user={user} selectedDate={dayjs(selectedDate).startOf('day').unix()}/>,
+            children: <BoughtOrdersTable user={user} selectedDate={""}/>,
           },
         ]}
       />
